requirements.txt
requests
icecream
loguru
tenacity
pandas
python-dotenv
cachier
scipy
seaborn
matplotlib
yfinance
